Revert "New Admin Account UI Refactor"
