+++ src/app/admin/(dashboard)/transactions/page.tsx
@@ -461,7 +461,7 @@ function TransactionForAccount() {
   );
 }
 
-export const tabs = [
+const tabs = [
   { value: "business-accounts", title: "Business Accounts" },
   { value: "issued-accounts", title: "Issued Accounts" },
   { value: "payout-accounts", title: "Payout Accounts" },
